Move socket tests to separate file
