+++ pkg/collector/corechecks/servicediscovery/module/socket_test.go
@@ -1,0 +1,102 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+package module
+
+import (
+	"net"
+	"os"
+	"strconv"
+	"strings"
+	"testing"
+
+	"github.com/shirou/gopsutil/v4/process"
+	"github.com/stretchr/testify/require"
+)
+
+func getSocketsOld(p *process.Process) ([]uint64, error) {
+	FDs, err := p.OpenFiles()
+	if err != nil {
+		return nil, err
+	}
+
+	// sockets have the following pattern "socket:[inode]"
+	var sockets []uint64
+	for _, fd := range FDs {
+		if strings.HasPrefix(fd.Path, prefix) {
+			inodeStr := strings.TrimPrefix(fd.Path[:len(fd.Path)-1], prefix)
+			sock, err := strconv.ParseUint(inodeStr, 10, 64)
+			if err != nil {
+				continue
+			}
+			sockets = append(sockets, sock)
+		}
+	}
+
+	return sockets, nil
+}
+
+const (
+	numberFDs = 100
+)
+
+func createFilesAndSockets(tb testing.TB) {
+	listeningSockets := make([]net.Listener, 0, numberFDs)
+	tb.Cleanup(func() {
+		for _, l := range listeningSockets {
+			l.Close()
+		}
+	})
+	for i := 0; i < numberFDs; i++ {
+		l, err := net.Listen("tcp", "localhost:0")
+		require.NoError(tb, err)
+		listeningSockets = append(listeningSockets, l)
+	}
+	regularFDs := make([]*os.File, 0, numberFDs)
+	tb.Cleanup(func() {
+		for _, f := range regularFDs {
+			f.Close()
+		}
+	})
+	for i := 0; i < numberFDs; i++ {
+		f, err := os.CreateTemp("", "")
+		require.NoError(tb, err)
+		regularFDs = append(regularFDs, f)
+	}
+}
+
+func TestGetSockets(t *testing.T) {
+	createFilesAndSockets(t)
+	p, err := process.NewProcess(int32(os.Getpid()))
+	require.NoError(t, err)
+
+	sockets, err := getSockets(p.Pid)
+	require.NoError(t, err)
+
+	sockets2, err := getSocketsOld(p)
+	require.NoError(t, err)
+
+	require.Equal(t, sockets, sockets2)
+}
+
+func BenchmarkGetSockets(b *testing.B) {
+	createFilesAndSockets(b)
+	b.ResetTimer()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		getSockets(int32(os.Getpid()))
+	}
+}
+
+func BenchmarkOldGetSockets(b *testing.B) {
+	createFilesAndSockets(b)
+	p, err := process.NewProcess(int32(os.Getpid()))
+	require.NoError(b, err)
+	b.ResetTimer()
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		getSocketsOld(p)
+	}
+}
